Moved LLCP link parameters configuration to NFC service initialization.
With the latest libnfc-nxp library, the LLCP socket internals are
initialized when configuring the LLCP link parameters because it
must be the first LLCP function called.

Due to this, these parameters are now set only once at the NFC
service startup instead of each peer discovery.

As a consequence, the NFC Manager initialization now use the
application properties (which contains link parameters) so these
properties are loaded before initialization.

Change-Id: I8e04c3135646a12a48c392d47cb47d09cd234ea1