Fixed presence checking in watchdog.
The presence check timer is not supposed to run during "other" operations
such as transceive. Previously, the presence check timer was simply reset.
If an operation would take longer than the presence check time to complete,
the presence check would run immediately after the operation has released
the concurrency lock. This is undesirable and not necessary - the app may
want to do a long series of transceives, in which case the presence check
gets in the way and delays the app.

This fix will pause() the watchdog completely, and only resume() when the
presence checking is done.

Change-Id: I06b1ab10c015861415cd91d6d5536d27145a885f