Support for multiple LLCP incoming sockets.
The global variable previously used to store a single incoming
connection request has been removed. Now, the requests are queued
in a list until the corresponding accept() method is called or
if the link goes down.

This should also fix issues related to multiple incoming connection
requests that can happen when activating/deactivating the link
fastly.

Change-Id: I3ab6c37b586b9f38b0de825e70f1efa06adf2062