Fix tag disconnect.
This patch fixes the tag disconnect sequence in 2 cases:

1.
When the NFC service disconnects a tag because it is not needed anymore, the
disconnection should be effective immediately and not delayed until the tag is
removed from the field. The NFC service is responsible for discovering again
the same tag if the polling loop is restarted.
The looping "presence check" in native code is then removed.

2.
Before the screen is turned off, the polling loop configuration and - if any
tag is present - the tag disconnection must be done accordingly to the "normal"
tag management. In particular the watchdog must be stopped properly.

Change-Id: I58ee27cede110c419052b4eef07f1e4bd35ae208