Fix isConnected() implementation.
The NFC API's isConnected() is mapped to isPresent on a NativeNfcTag
object, which in turn uses a cached value. This value was however not
updated properly when the internal presence check loop detected the
tag to be gone. The cases where isConnected() does work properly
are screen off and explicit disconnect.

Change-Id: Ieac76ceb1f161ead772ce9fd2acf273b23caca37