Improve DESFire check to determine NdefFormatable.
The previous code just checked for a response from the tag -
but some tags will just generate an error-response to the (custom)
DESFire getVersion command. These would be marked formatable incorrectly.

Additionaly now the length of the response is checked, and the status bytes
at the end of the response (which are fixed and well-known). Eventually
this code should live in libnfc.

Change-Id: Ib10cbc8f4e92977b66cf8353eb0a01b05c912c24