DO NOT MERGE. Set NFC EEPROM area to default value for FW 109.7.
This allows you to switch from FW 109.11 (ICS-MR1) and FW 109.7 (ICS-MR0).

This EEPROM location is used in ICS-MR1 for a new feature (Hybrid polling),
but the value used in MR1 causes polling to be disabled in MR0. EEPROM values
are stored on the NFC controller and will survive regular flashing. So the
result is that if you switch from MR1 to MR0 then polling will fail.

This fix is low risk, it just explicitly writes the value already stored in
this EEPROM location.

Bug: 5519093
Change-Id: I33d97134c68c03358f22b7773deea73c66166a98