Add watchdog to applyRouting().
We've seen some devices hang in doSelectSecureElement,
probably waiting for a response from the HW. Root cause
is unknown, may be due to a different access pattern
of Wallet or OTAProxy. Until we can reproduce better
and find the root cause, watchdog this method. Added
advantage is that we can have the relevant context in the logs
when the crash occurs.

Bug: 6585958
Change-Id: Ic797ea479869a68173fcd42049cc4ba8d009d2e0