Fix NFC low-power modes.
At the end of nfcManager_doInitialize(), enter low-power mode.
This logic is used after boot and before the user unlocks the screen.
If the user does not unlock the screen, the phone stays in low-power mode.

Additionally, only power off the chip if discovery is disabled
AND the SE is not selected.

Change-Id: I2bd2effce34b5dc586ad484c1da2ae86da49f0c1