Fast-fail SE open if it's activated in listen mode.
Users of the NFC-extras API usually want to talk
to the SE once it has completed a transaction.
The hard part is knowing when it is safe to connect
to the SE, as it will break the connection to any
reader. And when relinquishing the connection to
the SE, the reader may find the device again and
process another transaction.

This patch adds the following two conditions
for allowing to open the SE from the DH:

1) The SE may not be activated in listen mode

2) The RF field must have been off for at least 50 ms

Bug: 7275484
Change-Id: Ibde32a8e2aef045c17ab76ef08c72f96edfedaef