Don't hold a timed wakelock.
We used to acquire a timed wakelock when either resuming
or suspending. We don't want to take a timed wakelock
when suspending, since we want to suspend asap, but also
don't want to underlock the wakelock.

Instead of taking a timed wakelock, just abort
the process in case the NFC controller hangs.

Bug: 7341879
Change-Id: Ibd2f468264f60a39db3cb271a6f05c50dfb559bf