Reset RF field status on SE activation.
We've seen the NFCC get in a state where
it continuously reports field off/field on
events. The danger is that the device goes
to sleep right after a field on event. On
wake-up, the field off/field on loop is
gone, and the device prevents any access
to the SE from the DH because it thinks
the field is still present.

Bug: 7386840
Change-Id: I87b681124078fe69c0244efd7b14ebe3ffd1c161