Serialize applying card emulation route.
Instead of immediately applying the routing,
serialze the request with any outstanding
commands (including NFC enable/disable). This
prevents race conditions when NFC is being disabled
and the card emu routing gets changed at the same time.

Wait until the AsyncTask is complete, to make sure
that the routing is applied (if possible) by the time
the call returns.

Bug: 7418238
Change-Id: I7d92533179c02f4b6b01a86967737e64532317a1