Immediately connect to LLCP services.
When we discover a remote LLCP target, look at what
services the current foreground app has requested,
and immediately connect them.

This should increase our overall chance of success
of completing the transaction, as we no longer
need to setup the connection when the user taps
the screen to confirm the send.

It should also improve RIM interop; RIM devices
break the LLCP link if there are no service connections
within a certain time limit. Connecting the services
immediately prevents this.

Bug: 5155356
Change-Id: Ie4a578c1da31e3d5eb3b392b8236c681b45f9271