Improve Kovio reading reliability.
NfcBarcode tags such as Kovio are only activated very briefly,
just to spit out the uid. Consequent NDEF and reconnect() calls
by the NfcService will often fail, causing the tag to not be
reported properly to the upper layers.

For now, don't perform NDEF checking on these tags, and just
dispatch them when we see them.

Bug: 8576348
Change-Id: If011f6e99d85a61d19d9cea15719fab90cffad43