Android Beam direct invoke / share menu invoking.
This change allows Android Beam to be invoked
directly from an application.

Additionally, it adds handling for the ACTION_SEND
intent that is launched by various apps when using the
share menu. This will cause Android Beam to be shown
as an option in share menus. There's some logic needed
that translates the data in an ACTION_SEND intent to
either NDEF message or an Uri that can be shared using
handover.

Finally, the Beam UI code has been modified to instantiate
a proper window, in order for us to catch the back key
event to be able to back out of the Beam animation
(this wasn't possible previously).

Bug: 5134061
Change-Id: I9cbfc50109bb497973bae177c67d920f6f990789