Don't dispatch tags when NDEF read times out.
If we get a timeout status from the stack, store
that state so we don't dispatch tags with empty
NDEF messages with a read timeout. This was already
done when we timed out while reading NDEF meta-data,
but not the message itself.

Also, don't play sounds during detection at all;
NDEF messages are usually so small that the "start
detect" sound is immediately followed by either a
"success" sound or a "failure" sound; we now only
play a sound if the tag was successfully read and
will be dispatched.

Bug: 31165960
Change-Id: I4738731ad3b265df08a22666e52d801578dd329f